requirements.txt
pandas
numpy
matplotlib
plotly
scikit-learn
python-dotenv
minio
requests

streamlit
joblib
